chore(issuer): refactor how tier is removed with RemoveTierModal comp
