+++ packages/polymath-issuer/src/pages/sto/components/ConfigureSTOForm/forms/USDTieredSTO/RemoveTierModal.js
