--- packages/polymath-ui/src/components/ConfirmModal/ConfirmModal._mdx
@@ -1,17 +1,0 @@
----
-name: ConfirmModal
-menu: Modals
----
-
-import { Playground, PropsTable } from 'docz';
-import ConfirmModal from './';
-
-# Footer
-
-<PropsTable of={Footer} />
-
-## Basic usage
-
-<Playground>
-  <ConfirmModal />
-</Playground>
